move executor table related operations by a delegation class
This patch follows #1345 and continue refactor other methods in
`JdbcExecutorLoader`. This is the first step: We just move related
methods out of original class and keep its implementation without any
changes. Next, we will call new DB API to change its implementation.
