+++ azkaban-common/src/main/java/azkaban/executor/JdbcExecutorLoader.java
@@ -59,12 +59,16 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
   private static final Logger logger = Logger
       .getLogger(JdbcExecutorLoader.class);
   private final ExecutionFlowDBManager executionFlowDBManager;
+  private final ExecutorDBManager executorDBManager;
   private EncodingType defaultEncodingType = EncodingType.GZIP;
 
   @Inject
-  public JdbcExecutorLoader(final Props props, final CommonMetrics commonMetrics, final ExecutionFlowDBManager executionFlowDBManager) {
+  public JdbcExecutorLoader(final Props props, final CommonMetrics commonMetrics,
+                            final ExecutionFlowDBManager executionFlowDBManager,
+                            final ExecutorDBManager executorDBManager) {
     super(props, commonMetrics);
     this.executionFlowDBManager = executionFlowDBManager;
+    this.executorDBManager = executorDBManager;
   }
 
   public EncodingType getDefaultEncodingType() {
@@ -662,16 +666,7 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
    */
   @Override
   public List<Executor> fetchAllExecutors() throws ExecutorManagerException {
-    final QueryRunner runner = createQueryRunner();
-    final FetchExecutorHandler executorHandler = new FetchExecutorHandler();
-
-    try {
-      final List<Executor> executors =
-        runner.query(FetchExecutorHandler.FETCH_ALL_EXECUTORS, executorHandler);
-      return executors;
-    } catch (final Exception e) {
-      throw new ExecutorManagerException("Error fetching executors", e);
-    }
+    return this.executorDBManager.fetchAllExecutors();
   }
 
   /**
@@ -682,17 +677,7 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
    */
   @Override
   public List<Executor> fetchActiveExecutors() throws ExecutorManagerException {
-    final QueryRunner runner = createQueryRunner();
-    final FetchExecutorHandler executorHandler = new FetchExecutorHandler();
-
-    try {
-      final List<Executor> executors =
-        runner.query(FetchExecutorHandler.FETCH_ACTIVE_EXECUTORS,
-          executorHandler);
-      return executors;
-    } catch (final Exception e) {
-      throw new ExecutorManagerException("Error fetching active executors", e);
-    }
+    return this.executorDBManager.fetchActiveExecutors();
   }
 
   /**
@@ -703,22 +688,7 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
   @Override
   public Executor fetchExecutor(final String host, final int port)
     throws ExecutorManagerException {
-    final QueryRunner runner = createQueryRunner();
-    final FetchExecutorHandler executorHandler = new FetchExecutorHandler();
-
-    try {
-      final List<Executor> executors =
-        runner.query(FetchExecutorHandler.FETCH_EXECUTOR_BY_HOST_PORT,
-          executorHandler, host, port);
-      if (executors.isEmpty()) {
-        return null;
-      } else {
-        return executors.get(0);
-      }
-    } catch (final Exception e) {
-      throw new ExecutorManagerException(String.format(
-        "Error fetching executor %s:%d", host, port), e);
-    }
+    return this.executorDBManager.fetchExecutor(host, port);
   }
 
   /**
@@ -728,22 +698,7 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
    */
   @Override
   public Executor fetchExecutor(final int executorId) throws ExecutorManagerException {
-    final QueryRunner runner = createQueryRunner();
-    final FetchExecutorHandler executorHandler = new FetchExecutorHandler();
-
-    try {
-      final List<Executor> executors =
-        runner.query(FetchExecutorHandler.FETCH_EXECUTOR_BY_ID,
-          executorHandler, executorId);
-      if (executors.isEmpty()) {
-        return null;
-      } else {
-        return executors.get(0);
-      }
-    } catch (final Exception e) {
-      throw new ExecutorManagerException(String.format(
-        "Error fetching executor with id: %d", executorId), e);
-    }
+    return this.executorDBManager.fetchExecutor(executorId);
   }
 
   /**
@@ -779,30 +734,7 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
   @Override
   public Executor addExecutor(final String host, final int port)
     throws ExecutorManagerException {
-    // verify, if executor already exists
-    Executor executor = fetchExecutor(host, port);
-    if (executor != null) {
-      throw new ExecutorManagerException(String.format(
-        "Executor %s:%d already exist", host, port));
-    }
-    // add new executor
-    addExecutorHelper(host, port);
-    // fetch newly added executor
-    executor = fetchExecutor(host, port);
-
-    return executor;
-  }
-
-  private void addExecutorHelper(final String host, final int port)
-    throws ExecutorManagerException {
-    final String INSERT = "INSERT INTO executors (host, port) values (?,?)";
-    final QueryRunner runner = createQueryRunner();
-    try {
-      runner.update(INSERT, host, port);
-    } catch (final SQLException e) {
-      throw new ExecutorManagerException(String.format("Error adding %s:%d ",
-        host, port), e);
-    }
+    return this.executorDBManager.addExecutor(host, port);
   }
 
 
@@ -813,18 +745,7 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
    */
   @Override
   public void removeExecutor(final String host, final int port) throws ExecutorManagerException {
-    final String DELETE = "DELETE FROM executors WHERE host=? AND port=?";
-    final QueryRunner runner = createQueryRunner();
-    try {
-      final int rows = runner.update(DELETE, host, port);
-      if (rows == 0) {
-        throw new ExecutorManagerException("No executor with host, port :"
-            + "(" + host + "," + port + ")");
-      }
-    } catch (final SQLException e) {
-      throw new ExecutorManagerException("Error removing executor with host, port : "
-          + "(" + host + "," + port + ")", e);
-    }
+    this.executorDBManager.removeExecutor(host, port);
   }
 
   /**
@@ -917,21 +838,7 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
   @Override
   public Executor fetchExecutorByExecutionId(final int executionId)
     throws ExecutorManagerException {
-    final QueryRunner runner = createQueryRunner();
-    final FetchExecutorHandler executorHandler = new FetchExecutorHandler();
-    Executor executor = null;
-    try {
-      final List<Executor> executors =
-        runner.query(FetchExecutorHandler.FETCH_EXECUTION_EXECUTOR,
-          executorHandler, executionId);
-      if (executors.size() > 0) {
-        executor = executors.get(0);
-      }
-    } catch (final SQLException e) {
-      throw new ExecutorManagerException(
-        "Error fetching executor for exec_id : " + executionId, e);
-    }
-    return executor;
+    return this.executorDBManager.fetchExecutorByExecutionId(executionId);
   }
 
   @Override
@@ -1485,45 +1392,7 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
     }
   }
 
-  /**
-   * JDBC ResultSetHandler to fetch records from executors table
-   */
-  private static class FetchExecutorHandler implements
-    ResultSetHandler<List<Executor>> {
-    private static final String FETCH_ALL_EXECUTORS =
-      "SELECT id, host, port, active FROM executors";
-    private static final String FETCH_ACTIVE_EXECUTORS =
-      "SELECT id, host, port, active FROM executors where active=true";
-    private static final String FETCH_EXECUTOR_BY_ID =
-      "SELECT id, host, port, active FROM executors where id=?";
-    private static final String FETCH_EXECUTOR_BY_HOST_PORT =
-      "SELECT id, host, port, active FROM executors where host=? AND port=?";
-    private static final String FETCH_EXECUTION_EXECUTOR =
-      "SELECT ex.id, ex.host, ex.port, ex.active FROM "
-        + " executors ex INNER JOIN execution_flows ef "
-        + "on ex.id = ef.executor_id  where exec_id=?";
-
-    @Override
-    public List<Executor> handle(final ResultSet rs) throws SQLException {
-      if (!rs.next()) {
-        return Collections.<Executor> emptyList();
-      }
-
-      final List<Executor> executors = new ArrayList<>();
-      do {
-        final int id = rs.getInt(1);
-        final String host = rs.getString(2);
-        final int port = rs.getInt(3);
-        final boolean active = rs.getBoolean(4);
-        final Executor executor = new Executor(id, host, port, active);
-        executors.add(executor);
-      } while (rs.next());
-
-      return executors;
-    }
-  }
-
-  /**
+ /**
    * JDBC ResultSetHandler to fetch records from executor_events table
    */
   private static class ExecutorLogsResultHandler implements
